Add 'get_recursive' option for legacy whois to RIPE NCC
This patch adds an argument `get_recursive` to `lookup_whois()`. The
argument default is True, which causes no change to the behavior. If
the user passes `get_recursive=False`, and the IP is allocated to
RIPE, this causes the `-r` flag to be included in the query string.
The `-r` flag asks RIPE WHOIS servers not to perform recursive
queries, such as for related "person" or "role" objects. This helps
users avoid IP blacklisting by RIPE.

The `-r` flag is only supported by RIPE, so this patch only affects
queries to RIPE servers.